+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Available types:
 
 ### Added
 - CLI option "-f|--failfast", it stops execution of tests on first error or failure.
+- CLI option "-s| --showtags", shows the tags for all tests, does not run the tests.
 
 ### Changed
 - Traceback when error occures in test is now printed with the docstring from the test.