+++ CHANGELOG.md
@@ -15,6 +15,9 @@ Available types:
 
 ## [Unreleased]
 
+
+
+## [2.3.0]
 ### Added
 - CLI option "-f|--failfast", it stops execution of tests on first error or failure.
 - CLI option "-s| --showtags", shows the tags for all tests, does not run the tests.